DBZ-8586 Implement Wasm accessors to all the Struct data types

+++ debezium-scripting/debezium-scripting/src/test/resources/wasm/go/filter1.go
@@ -7,7 +7,7 @@ import (
 //export process
 func process(proxyPtr uint32) uint32 {
 	var op = debezium.GetString(debezium.Get(proxyPtr, "value.op"))
-	var beforeId = debezium.GetInt(debezium.Get(proxyPtr, "value.before.id"))
+	var beforeId = debezium.GetInt8(debezium.Get(proxyPtr, "value.before.id"))
 
 	// value.op != 'd' || value.before.id != 2
 	return debezium.SetBool(op != "d" || beforeId != 2)
